refactor: WI-0304 payroll service modular split phase5
