+++ ROADMAP.md
@@ -574,3 +574,4 @@ Phase 8: Extensions (ATS, performance, expenses, analytics)
 - WI-0301 employee/admin locale dynamic UI gap fix phase2(/employee locale-aware status/badge/work-type/log labels + leave-type display harmonization + /admin invite role/delivery locale labels + WI-0301 regression test)
 - WI-0302 payroll decomposition phase4 baseline(src/features/payroll/year-end-filing-lifecycle-helpers.ts extraction for filing submission summary/timeline builders + src/features/payroll/service.ts lifecycle delegation rewiring + WI-0302 regression test)
 - WI-0303 admin/employee locale dynamic UI gap fix phase3(`/admin`/`/employee` mixed-language hardcoded label cleanup + resubmit aria-label mojibake(`???`) fix + locale-aware runtime copy refinement + WI-0303 regression test)
+- WI-0304 payroll decomposition phase5 baseline(src/features/payroll/year-end-filing-ack-catalog-helpers.ts extraction for filing ack catalog constants/payload resolver + src/features/payroll/service.ts delegation rewiring + WI-0304 regression test)
